NO-ISSUE: Move exit error to separate package
This patch moves the `internal.ExitError` to a new `exit` package. That
makes it more natural to use.
