+++ ztp/internal/cmd/dev/setup/dev_setup_cmd.go
@@ -20,6 +20,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal"
+	"github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/exit"
 	"github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/labels"
 )
 
@@ -84,7 +85,7 @@ func (c *Command) run(cmd *cobra.Command, argv []string) (err error) {
 			"Failed to create applier: %v\n",
 			err,
 		)
-		return internal.ExitError(1)
+		return exit.Error(1)
 	}
 	err = applier.Apply(ctx, nil)
 	if err != nil {
@@ -93,7 +94,7 @@ func (c *Command) run(cmd *cobra.Command, argv []string) (err error) {
 			"Failed to apply objects: %v\n",
 			err,
 		)
-		return internal.ExitError(1)
+		return exit.Error(1)
 	}
 
 	return nil
