+++ ztp/internal/exit/error.go
@@ -12,21 +12,21 @@ implied. See the License for the specific language governing permissions and lim
 License.
 */
 
-package internal
+package exit
 
 import "fmt"
 
-// ExitError is an error type that contains a process exit code. This is itended for situations
-// where you want to call os.Exit only in one place, but also want some deeply nested functions to
-// decide what should be the exit code.
-type ExitError int
+// Error is an error type that contains a process exit code. This is itended for situations where
+// you want to call os.Exit only in one place, but also want some deeply nested functions to decide
+// what should be the exit code.
+type Error int
 
 // Error is the implementation of the error interface.
-func (e ExitError) Error() string {
+func (e Error) Error() string {
 	return fmt.Sprintf("%d", e)
 }
 
 // Code returns the exit code.
-func (e ExitError) Code() int {
+func (e Error) Code() int {
 	return int(e)
 }
